Refactored the GLSL program interface to accommodate threaded rendering
Refactored World Wind's GpuProgram and associated methods on GpuObject cache to separate program construction and GLSL compilation/linking. This enables identification of the desired GLSL program on the Activity thread, and compilation/linking on the OpenGL thread.

Replaced the existing pattern for GpuProgram construction and cache access with a more extensible one. The previous pattern made it difficult to add new GLSL programs and required that each GLSL program be addressed in the GPU cache using a class name. The latter issue had the effect of requiring rendering code to create a new program class in order to create a new instance in OpenGL.